split widgets, update exports
