+++ src/browser.ts
@@ -1,5 +1,4 @@
-export * as Controllers from './widgets/controllers/index.js';
 export * as Models from './widgets/models/index.js';
-export * as Views from './widgets/views/index.js';
 export * as OpsTypes from './ops-types.js';
+export * from './widgets/views/index.js';
 export { AwsUtilizationProvider } from './aws-utilization-provider.js';
