+++ src/widgets/views/components/index.ts
@@ -1,0 +1,9 @@
+export * from './confirm-recommendations.js';
+export * from './confirm-single-recommendation.js';
+export * from './recommendation-overview.js';
+export * from './recommendations-action-summary.js';
+export * from './recommendations-table.js';
+export * from './service-table-row.js';
+export * from './side-panel-related-resources.js';
+export * from './side-panel-usage.js';
+export * from './utilization-recommendations-ui.js';
