+++ src/widgets/views/components/recommendations-action-summary.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { Box, Button, Flex, Heading, Icon, Spacer, Stack, Text } from '@chakra-ui/react';
 import { DeleteIcon, ArrowForwardIcon, ArrowDownIcon } from '@chakra-ui/icons';
 import { TbVectorBezier2 } from 'react-icons/tb/index.js';
-import { filterUtilizationForActionType, getNumberOfResourcesFromFilteredActions } from '../../utils/utilization.js';
-import { ActionType } from '../../types/types.js';
-import { RecommendationsActionSummaryProps } from '../../types/utilization-recommendations-types.js';
+import { filterUtilizationForActionType, getNumberOfResourcesFromFilteredActions } from '../../../utils/utilization.js';
+import { ActionType } from '../../../types/types.js';
+import { RecommendationsActionSummaryProps } from '../../../types/utilization-recommendations-types.js';
 import { TbRefresh } from 'react-icons/tb/index.js';
