+++ src/widgets/views/components/utilization-recommendations-ui.tsx
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
-import { ActionType } from '../../types/types.js';
+import { ActionType } from '../../../types/types.js';
 import { RecommendationsActionSummary } from './recommendations-action-summary.js';
 import { RecommendationsTable } from './recommendations-table.js';
 import { ConfirmRecommendations } from './confirm-recommendations.js';
-import { UtilizationRecommendationsUiProps } from '../../types/utilization-recommendations-types.js';
+import { UtilizationRecommendationsUiProps } from '../../../types/utilization-recommendations-types.js';
 
 enum WizardSteps {
   SUMMARY='summary',
