+++ src/widgets/views/index.ts
@@ -1,0 +1,2 @@
+export * from './aws-utilization-recommendations.js';
+export * from './aws-utilization.js';
